fetchWordのシフト時の不具合修正
uint8のまま8bitシフトしてたため値がゼロになってしまった。そのあとにuint16してもゼロのまま。
シフトする前にupperをuint16にしておく。

+++ cpu/fetch.go
@@ -13,5 +13,8 @@ func (cpu *Cpu) fetchWord() uint16 {
 	upper := cpu.CpuBus.ReadByCpu(cpu.Registers.GetPc())
 	cpu.Registers.IncrementPc()
 
-	return uint16(upper << 8 | lower)
+
+	// upper, lower共に uint16にしておかないといけない
+	// upperで8bitシフトするのでuint16の型にしてからシフトしないと0になるため
+	return uint16(upper) << 8 | uint16(lower)
 }
